Use fns instead of relying on named! macro (#1)
This will help make our life easier with recursion and such. Most of the struggle is figuring out the return types of the different constructs `nom` gives us. My understanding is that, when you have a macro named `my_macro`, say, and it takes a single argument and returns a function, doing `my_macro!(some_arg)` both _expands_ the macro _and_ calls the resulting function it produces. However, doing `my_macro(some_arg)` (without the `bang!`) simply returns the function it produces, meaning something like this `my_macro(some_arg)(some_other_arg)` could make sense (and that's how this PR is using `take_while` among others).

EDIT: This style is very reminiscent of `do` notation in Haskell. The `?` operator suggested by the God @gabaconrado gives us a short circuiting way of piping the the result of parsers to the next ones.

The next step is to implement variadic arguments and expressions as operators (the first argument to a list doesn't have to be a `char`, it can be an entire expression). Also, not every atom is a Num, so we'll make that into a rich enum as well.