Implement actual expression struct and evaluation (#2)
Make the expression struct account for an expression as its first element, and allow for lists of arbitrary length. Remember that expressions in Lisp are heterogenous lists. Evaluating expressions always amounts to: evaluate the first element, then evaluate all remaining elements, then apply the operation the first element represents to all other elements.

At the moment in time of this application, all elements should have the form of a "finished" argument. They can still be lists, but in that case, for the function call to make sense, it should be a function that takes lists as a parameter somehow.

The first element, after evaluation, should always resolve to a function, if the top-level list is being `eval`ed. Otherwise, a runtime error should occur.