Built-in arithmetic functions and evaluation errors (#5)
Basically chapters 7 through 9 of buildyourownlisp.com.

We handle evaluation error in a hopefully parametrizable EvalResult. We say parametrizable because we know the kind of expressions certain built-in functions return: arithmetic returns numbers for instance, and can only result in a certain number of errors. One of the TODOs is about this.

The structure for all evaluation errors is within this PR, and also for all future built-in functionality (of which we really only plan to implement `def` and `fn`).