Def, local vars/environments and let bindings (#8)
Lots of stuff again. "built_in" module now exports an fn to create a new environment from scratch, which main uses (main is where the "REPL" of this little thing is running). Environment also is now split into two values: the original struct, and a linked list of HashMaps which contain local bindings. The there is that this implements shadowing.

For functions: whenever a function is called, its arguments are bound to the parameter names, overwriting any global defs or local vars defined in a scope "above". This is expressed by adding a head to the list, and if any inner scope expressions looks up that symbol, it's guaranteed to find the function's bindings before the upper level's ones. The way we were doing it before is just copying the entire environment any time a function was called, which is pretty redundant.

For let bindings: pretty much the same, except the names in a let binding are substituted in the same scope as the values the should hold, unlike functions, whose evaluation happens at the caller, before "going in" to the function's environment. But the basic idea is the same: add a new node to the linked list and pass that down to the body of the binding.

This is very fun c: